1.3.4 ヨシ！

+++ src/main/java/be4rjp/sclat/Sclat.java
@@ -117,7 +117,7 @@ public class Sclat {
     
     public static boolean isNumber(String s) {
         try {
-            Integer.parseInt(s);
+            Double.parseDouble(s);
             return true;
         } catch (NumberFormatException e) {
             return false;
